THIS is MADNess
I am not getting anywhere with this code, I am trying to create a list
of questions. Then those questions are to be clickable so that they
open up a new window to display my answers, but its just not coming
together. I keep getting a parse error with an unexpected token at the
end of the code..